--- Assets/Prefabs/Main Camera.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 318cb464875a448208c2f059cc92406f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
