+++ Assets/Scripts/Camera/CameraCollider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c27e298ec92d74192aa231725dbdf0e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
