+++ Assets/Scripts/Camera/CameraWeapon.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e5b10e0ab80ef447c8912f20db44a08a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
