+++ Assets/Scripts/Event.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76175cdba0f8d40cc95289e55df283ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
