+++ Assets/Scripts/Event/CameraEventController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 21b735b14902c4704a2469379ff55380
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
