--- Assets/Scripts/EventController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e95b4149867664ba3b5bcc02dc563cdf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
