--- Assets/Scripts/Light Hacking.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5ce3abfc39c5e07469db8df375f226b4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
