--- Assets/Scripts/TrainingDummy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9889025783c524b0e8073f8c9aeb949e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
