+++ Assets/Scripts/UI/DeathScreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 63b502d4436ed4a62894747577c5a4d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
