+++ Assets/Scripts/UI/Poloroid.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de49a475742e24a0390b40044fb2c7e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
