+++ Assets/Sprites/Poloroid.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 900d5497025b144e580ce41fccc3bf59
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
