+++ Assets/Sprites/Poloroid/Poloroid.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e303a9c4c4b7b44f68bcdf82d10169d8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8600000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
