+++ Assets/Sprites/Poloroid/Poloroid.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0640b9540f6644a55a1b47d5021a70ef
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
