+++ Assets/Sprites/Poloroid/Poloroid.shader.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: c78ee7e6e59ad4277a7193aa758e7e89
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures:
+  - _MainTex: {instanceID: 0}
+  - _SubTex: {fileID: 8600000, guid: e303a9c4c4b7b44f68bcdf82d10169d8, type: 2}
+  nonModifiableTextures: []
+  preprocessorOverride: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
